Modification de la navbar pour qu'elle soit responsive et modification du style pour le nom de famille
