+++ frontend/src/components/Desktop/Navbar/Navbar.scss
@@ -1,0 +1,159 @@
+/* CustomNavbar.scss */
+
+.navbar {
+  width: 100vw;
+  height: 6rem;
+  background-color: #10131e;
+  position: fixed;
+  border-bottom: 1px solid #594848;
+
+  .navbarContent {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+    height: 100%;
+    // margin: 0 auto;
+
+    .leftContainer {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      margin-left: 4rem;
+      .familyName {
+        font-size: 3rem;
+        text-decoration: none;
+        font-weight: bold;
+        color: transparent;
+        margin-right: 2rem;
+        margin-left: 1rem;
+        border-radius: 20px;
+
+        /* Ajout d'un effet de néon avec un box-shadow */
+        text-shadow: 
+          2px 2px 0 #2aa6ff, 
+          -2px -2px 0 #10131e, 
+          2px -2px 0 #ffffff, 
+          -2px 2px 0 #2aa6ff,
+          2px 0 0 #2aa6ff,
+          0 2px 0 #10131e,
+          -2px 0 0 #ffffff,
+          0 -2px 0 #2aa6ff;
+      
+
+       
+    
+
+        &:hover {
+          text-decoration: none;
+        }
+      }
+    }
+    .centerContainer {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      margin-left: 2rem;
+      width: 100%;
+      .searchContainer {
+        background-color: #10131e;
+        max-width:30rem;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        border: 1px solid #fff;
+        border-radius: 5px;
+        padding: 0.5rem;
+
+        .searchIcon {
+          display: flex;
+          align-items: center;
+          font-size: 1.2rem;
+          margin-right: 0.5rem;
+          color: #2aa6ff;
+        }
+
+        .searchInput {
+          width: 10rem;
+          background-color: transparent;
+          border: none !important;
+          outline: none !important; ;
+          color: #fff;
+          &::placeholder {
+            color: #ccc;
+          }
+        }
+
+        .searchButton {
+          width:5rem; 
+          background-color: #10131e;
+          color: #fff;
+          border: 1px solid #fff;
+          border-radius: 5px;
+          padding: 0.25rem 0.5rem;
+          font-weight: bold;
+          cursor: pointer;
+          transition: all 0.5s;
+          font-size: 12px;
+          margin-right: 1rem;
+
+          &:hover {
+            background-color: #009ea5;
+          }
+        }
+      }
+    }
+
+    .rightContainer {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      margin-left: 2rem;
+      // width: 20%;
+
+      .navbarIcon {
+        display: flex;
+        align-items: center;
+
+        
+
+          .icon {
+            margin-right: 2rem;
+            font-size: 1.5rem;
+            color: #2aa6ff;
+            transition: color 0.2s;
+
+            &:hover {
+              color: #009ea5;
+            }
+          }
+        }
+      
+
+      .userAvatarContainer {
+        width: 30%;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        .imgAvatar {
+          width: 3.5rem;
+          height: 3.5rem;
+          border: 2px solid #2aa6ff;
+          border-radius: 50%;
+          margin-right: 1rem;
+          cursor: pointer;
+          transition: border-color 0.2s;
+
+        }
+      }
+    }
+  }
+
+  // Utilisez une Media Query pour rendre l'image de profil plus grande sur les écrans plus grands
+  @media screen and (max-width: 768px) {
+    .user-image-container .user-image {
+      width: 1.5rem;
+      height: 1.5rem;
+    }
+  }
+}
