--- frontend/src/components/navbar/Navbar.scss
@@ -1,104 +1,0 @@
-@import "../../style/variables.scss";
-
-.navbar {
-  height: 8rem;
-  border-bottom: 1px solid #594848;
-  color: $primary-color;
-  position: sticky;
-  background-color: $background-fond;
-}
-
-.navContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  
-  .familyContainer {
-    width: 20rem;
-    height: 5rem;
-    
-    .familyName {
-      color: #fff;
-      font-size: 4rem;
-      background: linear-gradient(
-        to right,
-        rgb(10, 54, 248),
-        rgb(9, 234, 174)
-        );
-        -webkit-background-clip: text;
-        background-clip: text;
-    color: transparent;
-    text-align: center;
-  }
-}
-
-
-.searchContainer {
-  width: 30rem !important;
-  display: flex;
-  align-items: center;
-  border: 1px solid #594848;
-  border-radius: 15px;
-  margin: 1rem;
-  padding: 0.5rem;
-  
-  .searchIcon {
-    font-size: 1.5rem;
-    padding-left: 0.5rem;
-  }
-  
-  .searchInput {
-    width: 1rem !important;
-    background-color: transparent;
-    border: none !important;
-    outline: none !important;
-    color: #2aa6ff;
-    
-    &::placeholder {
-      border: none !important;
-      outline: none !important;
-      color: #2aa6ff;
-    }
-  }
-  button{
-    width: 5rem;
-    background-color: transparent;
-    border-radius: 10px !important;
-    font-size: 12px;
-  }
-}
-
-.icon {
-  color: #2aa6ff;
-  font-size: 2rem;
-  margin-left: 3rem;
-  cursor: pointer;
-  
-  &:hover {
-    color: $secondary-color;
-  }
-}
-
-.tooltip-text {
-  font-size: 1rem;
-  color: $primary-color !important;
-  font-weight: 500;
-}
-
-.imgUser {
-  width: 3rem !important;
-  margin-left: 3rem;
-  border: 1px solid #594848;
-      object-fit: cover;
-    }
-    
-    
-    
-    
-    // Supprimer la navbar en version mobile et tablette
-    @media screen and (max-width: 768px) {
-      .navbar {
-        display: none;
-      }
-    }
-  }
